requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
dnspython==2.8.0
fake-useragent==2.0.3
Flask==3.1.2
fonttools==4.55.3
google==3.0.0
google-auth==2.41.1
google-genai==1.41.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pymongo==4.15.2
pyparsing==3.2.0
python-csv==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
requests==2.32.5
rsa==4.9.1
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
xlrd==2.0.1
zope.interface==7.2
